+++ guillotina/schema/_bootstrapfields.py
@@ -103,13 +103,14 @@ class Field(Attribute):
     interface = None
     _Element__tagged_values = None
     _validators = None
+    _required = True
 
     def __init__(
         self,
         title="",
         description="",
         __name__="",
-        required=True,
+        required=None,
         readonly=False,
         constraint=None,
         default=None,
@@ -154,7 +155,6 @@ class Field(Attribute):
         super(Field, self).__init__(__name__, __doc__)
         self.title = title
         self.description = description
-        self.required = required
         self.readonly = readonly
         if constraint is not None:
             self.constraint = constraint
@@ -169,6 +169,13 @@ class Field(Attribute):
 
         if missing_value is not self.__missing_value_marker:
             self.missing_value = missing_value
+
+        if required is not None:
+            self._required = required
+
+    @property
+    def required(self):
+        return self._required
 
     def validator(self, func):
         if self._validators is None: